+++ CHANGELOG.md
@@ -8,13 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Removed
 
-- Dropped support for Python 3.8 or older. (Requires Python >= 3.9)
+- Dropped support for Python 3.8 or older. (Requires Python >= 3.10)
 
 ### Core
 
 #### Added
 
 - SSL torrents support for secure peer-to-peer connections. See [libtorrent docs](https://libtorrent.org/manual-ref.html#ssl-torrents) for further implementation details.
+- Add option to announce to trackers in all tiers (uTorrent behavior). (#1395)
 
 #### Changed